[FIX] 워크스루 플로우 오류 수정
