+++ Projects/Presentation/Presentation/Sources/Assembly/RecordAssembly.swift
@@ -1,0 +1,18 @@
+//
+//  RecordAssembly.swift
+//  Presentation
+//
+//  Created by dong eun shin on 7/5/25.
+//
+
+import UIKit
+import Swinject
+import SwinjectAutoregistration
+
+public final class RecordAssembly: Assembly {
+  public init() {}
+
+  public func assemble(container: Container) {
+
+  }
+}
