+++ Projects/Presentation/Presentation/Sources/MainTabBar/Record/RecordDetail/Cell/RecordDetailAchievementTableCell.swift
@@ -41,7 +41,7 @@ public class RecordDetailAchievementTableCell: BaseTableViewCell {
   public override func initUI() {
     super.initUI()
     
-    contentView.backgroundColor = FRColor.Fg.Nuetral.gray0
+    contentView.backgroundColor = .clear
 
     contentView.addSubview(rootStack)
     rootStack.snp.makeConstraints {
