+++ Projects/Presentation/Presentation/Sources/MainTabBar/Record/RecordDetail/Cell/RecordDetailRecordTableCell.swift
@@ -96,7 +96,7 @@ public class RecordDetailRecordTableCell: BaseTableViewCell {
   public override func initUI() {
     super.initUI()
     
-    contentView.backgroundColor = FRColor.Fg.Nuetral.gray0
+    contentView.backgroundColor = .clear
     contentView.addSubview(rootStack)
     
     rootStack.snp.makeConstraints {
