+++ Projects/Presentation/Presentation/Sources/MainTabBar/Record/RecordDetail/Cell/RecordDetailTitleTableCell.swift
@@ -55,7 +55,7 @@ public class RecordDetailTitleTableCell: BaseTableViewCell {
   public override func initUI() {
     super.initUI()
     
-    contentView.backgroundColor = FRColor.Fg.Nuetral.gray0
+    contentView.backgroundColor = .clear
 
     contentView.addSubview(rootStack)
     rootStack.snp.makeConstraints {
